initial test/implementation for updating blogs 📜
added test to make sure the version is incremented when an update is made and created a test to ensure that the content could be changed as well

added test/implementation for updating blogs

The user should be allowed to make certain changes. No user should be able to create a blog and change the author. Thats fake news. (Make PHP great again)